# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewFlow Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Users ───
...
  /analytics/dashboard:
    get:
      operationId: getDashboardStats
      tags: [analytics]
      summary: Get dashboard KPI stats
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, custom]
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /analytics/reviews-over-time:
    get:
      operationId: getReviewsOverTime
      tags: [analytics]
      summary: Get reviews chart data over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews over time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSeriesPoint"
  /analytics/scans-over-time:
    get:
      operationId: getScansOverTime
      tags: [analytics]
      summary: Get QR scans chart data over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Scans over time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSeriesPoint"
  /analytics/top-clients:
    get:
      operationId: getTopClients
      tags: [analytics]
      summary: Get top performing clients
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopClient"
  /analytics/rating-distribution:
    get:
      operationId: getRatingDistribution
      tags: [analytics]
      summary: Get rating distribution breakdown
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Rating distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingDistributionItem"
  /analytics/revenue-over-time:
    get:
      operationId: getRevenueOverTime
      tags: [analytics]
      summary: Get monthly revenue chart data
      parameters:
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue over time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePoint"

  # ─── Audit Logs ───
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── Auth ───
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, client]
        status:
          type: string
          enum: [active, inactive]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, status, createdAt, updatedAt]
...
    DashboardStats:
      type: object
      properties:
        totalQrScans:
          type: integer
        totalReviews:
          type: integer
        totalGoogleRedirects:
          type: integer
        positiveReviews:
          type: integer
        negativeReviews:
          type: integer
        conversionRate:
          type: number
        activeClients:
          type: integer
        inactiveClients:
          type: integer
        monthlyRevenue:
          type: number
        expiringSubscriptions:
          type: integer
        pendingInvoices:
          type: integer
        scansGrowth:
          type: number
        reviewsGrowth:
          type: number
        revenueGrowth:
          type: number
      required: [totalQrScans, totalReviews, totalGoogleRedirects, positiveReviews, negativeReviews, conversionRate, activeClients, inactiveClients, monthlyRevenue, expiringSubscriptions, pendingInvoices, scansGrowth, reviewsGrowth, revenueGrowth]
    TimeSeriesPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: integer
        label:
          type: ["string", "null"]
      required: [date, value]
    TopClient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessName:
          type: string
        totalScans:
          type: integer
        totalReviews:
          type: integer
        conversionRate:
          type: number
        positiveRate:
          type: number
      required: [id, name, businessName, totalScans, totalReviews, conversionRate, positiveRate]
    RatingDistributionItem:
      type: object
      properties:
        rating:
          type: integer
        count:
          type: integer
        percentage:
          type: number
      required: [rating, count, percentage]
    RevenuePoint:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        invoicesCount:
          type: integer
      required: [month, revenue, invoicesCount]

    # ─── Audit Log ───
...